fix float precision

+++ models/expense_model.go
@@ -3,5 +3,5 @@ package models
 type Expense struct {
 	ID       string  `json:"id"`
 	Category string  `json:"category"`
-	Price    float64 `json:"price"`
+	Price    float32 `json:"price"`
 }
